+++ CHANGES.md
@@ -1,5 +1,12 @@
 # CHANGES for integrity-matters
 
+## 0.2.0
+
+- Fix: Avoid using globs when `outputPath` is set and need for `noGlobs`
+    when using `outputPath`; print proper path
+- Enhancement: Log file path when finished writing file
+- Enhancement: Deterministic logging order
+
 ## 0.1.0
 
 - Initial version